+++ PaymentSystem/PaymentSystem.Domain/BankAccount.cs
@@ -11,7 +11,7 @@ namespace PaymentSystem.Domain
 
         public DateTime DateCreated { get; set; }
 
-        public decimal AmountOfMoney { get; set; }
+        public double AmountOfMoney { get; set; }
 
         public IEnumerable<Card> PaymentCards { get; set; }
     }
